% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{flexfilter}
\alias{flexfilter}
\alias{flexfilterUI}
\alias{flexfilter_server}
\title{Flex filter}
\usage{
flexfilterUI(id, btn_class = "")

flexfilter_server(id, data, variables_only = FALSE, search_threshold = 20L)
}
\arguments{
\item{id}{Module id.}

\item{btn_class}{Additional classes to pass to the button}

\item{data}{Dataset.}

\item{variables_only}{Set to \code{TRUE} to not display filters and only display
variable names.}

\item{search_threshold}{if the number of variables available exceed the \code{search_threshold}
then a search box appears.}
}
\description{
Flex filter component.
}
