% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{flexfilter}
\alias{flexfilter}
\alias{flexfilterUI}
\alias{flexfilter_server}
\title{Flex filter}
\usage{
flexfilterUI(id)

flexfilter_server(id, data)
}
\arguments{
\item{id}{Module id.}

\item{data}{Dataset.}
}
\description{
Flex filter component.
}
